Brings in the large upstream lint-hardening/ownership/bounds batch
(~300 commits: token-typed ownership migration, explicit scan/loop
bounds, CI matrix fixes, Windows ARM fixes) alongside this branch's
POSIX-utility Tier 0-3 work. Only Makefile, src/internal/util.h and
src/sh/builtin.c touch the same files on both sides.

+++ fuzz/linux_pilot_test_open.c
@@ -21,8 +21,8 @@
 
 extern long syscall(long number, ...);
 extern int printf(const char *, ...);
-extern void *memset(void *, int, unsigned long);
-extern int memcmp(const void *, const void *, unsigned long);
+extern void *memset(void *, int, size_t);
+extern int memcmp(const void *, const void *, size_t);
 
 #define SYS_unlinkat 35
 #define SYS_mkdirat  34
